Add support of FreeBSD 13.0.
The interfaces that limit cap_sysctl MIBs has been changed since FreeBSD 13.0.
Now configure script checks if 'cap_sysctl_limit_name' function is available.
If it is available, asmctl use the function.
The old interfaces are used when 'cap_sysctl_limit_name' is not available.

Fix to cache catpages for printing an error message by 'strerror'.
It was caused capability violation because NLS database can not be opened in
capability mode.

Fix to handle 'prefix' variable correctly in configure script.