Fix issue where asmctl would not dim backlight completely if backlight=2.
If backlight=2, asmctl will set backlight=1. However, the backlight may truly
be set to 0 due to a bug in the screen. Although backlight will be set to 0,
the screen will not be completely off as it should be. This patch ensures that
if backlight=2, it will pretend that backlight=1; meaning backlight=0 will be
set correctly.

Also fix indentation and comment.